Update some tables and add checking on the not null values of a row
Add the time recall column in the chats table. This value specify to the bot when the bot must send the recall to the user. Also this value has the default value.

Edit the notes table. The id of each note is unnecessary. Instead of that the primary key was changed and the name of a note was added.