Update the tables of the database
Before the chat_id was containing in the Chats table. Now it will be contained in the Conversations table and the other tables will reference on it. It was changed because of the Conversations table are used before the other tables. Since the ConversationHandler calls handlers that call the getting chat_id method, so it was decided to make the Conversation table main.